Split editor action validation
